+++ src/contexts/SharedInfraContext.ts
@@ -1,7 +1,6 @@
 import { createContext } from 'react';
 
 const defaultSharedInfra: SharedInfraServices = {
-  lookupCacheService: undefined as unknown as ILookupCacheService,
   commonStatusService: undefined as unknown as ICommonStatus,
 };