+++ src/contexts/index.ts
@@ -1,3 +1,2 @@
-export { ServicesContext } from './ServicesContext';
 export { SharedInfraContext } from './SharedInfraContext';
 export { SchemaPipelineContext } from './SchemaPipelineContext';