+++ src/features/resources/hooks/useProcessedRecordAndSchema.test.ts
@@ -1,4 +1,4 @@
-import { recordToSchemaMappingService } from '@/test/__mocks__/common/hooks/useServicesContext.mock';
+import { recordToSchemaMappingService } from '@/test/__mocks__/common/hooks/useSchemaPipeline.mock';
 import { setInitialGlobalState } from '@/test/__mocks__/store';
 
 import { act, renderHook } from '@testing-library/react';