+++ src/features/resources/hooks/useProcessedRecordAndSchema.ts
@@ -5,7 +5,7 @@ import { DUPLICATE_RESOURCE_TEMPLATE } from '@/common/constants/resourceTemplate
 import { StatusType } from '@/common/constants/status.constants';
 import { getAdjustedRecordContents, wrapRecordValuesWithCommonContainer } from '@/common/helpers/record.helper';
 import { applyIntlToTemplates } from '@/common/helpers/recordFormatting.helper';
-import { useServicesContext } from '@/common/hooks/useServicesContext';
+import { useSchemaPipeline } from '@/common/hooks/useSchemaPipeline';
 import { logger } from '@/common/services/logger';
 import { UserNotificationFactory } from '@/common/services/userNotification';
 
@@ -25,7 +25,7 @@ export const useProcessedRecordAndSchema = () => {
   const { setRecord } = useInputsState(['setRecord']);
   const { formatMessage } = useIntl();
   const { userValuesService, schemaWithDuplicatesService, recordNormalizingService, recordToSchemaMappingService } =
-    useServicesContext() as Required<ServicesParams>;
+    useSchemaPipeline();
 
   const getProcessedRecordAndSchema = useCallback(
     async ({