+++ src/features/search/ui/hooks/useHubSearchPreviewQuery.test.tsx
@@ -1,4 +1,4 @@
-import '@/test/__mocks__/common/hooks/useServicesContext.mock';
+import '@/test/__mocks__/common/hooks/useSchemaPipeline.mock';
 import { setInitialGlobalState } from '@/test/__mocks__/store';
 
 import { renderHook } from '@testing-library/react';