+++ src/providers/index.ts
@@ -1,4 +1,3 @@
 export { ServicesProvider } from './ServicesProvider';
 export { AsyncIntlProvider } from './AsyncIntlProvider';
 export { QueryProvider } from './QueryProvider';
-export { SchemaPipelineProvider } from './SchemaPipelineProvider';