+++ src/types/simpleLookup.d.ts
@@ -1,5 +1,0 @@
-interface ILookupCacheService {
-  save: (key: string, data: MultiselectOption[]) => void;
-  getAll: () => Record<string, MultiselectOption[]>;
-  getById: (id: string) => MultiselectOption[];
-}