+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Refactor record control hooks. Refs [UILD-744].
 * Refactor resource processing and preview loading. Refs [UILD-744].
 * Refactor resource loading. Refs [UILD-816].
+* Refactor services providers and simple lookup data loading. Refs [UILD-816].
 
 [UILD-744]:https://folio-org.atlassian.net/browse/UILD-744
 [UILD-816]:https://folio-org.atlassian.net/browse/UILD-816